Fix barchart series displaying with incorrect padding
Because setYScales() is called multiple times, the barchart-specific
change to axis.domain was reused in a later call when we checked to see
if axis.domain was already computed and updated the existing
scale.domain instead of recreating a new scale.domain. Because we
always create a new scale.domain now, we need to ensure that the
axis.domain is correct before creating the scale.domain. Moving the
change forward resolved the display padding issue.

Also fixes up some tests.